Kenzie/Sub-MAF-09: Added preference manager values and updated bug on safety and operation param card
