+++ app/src/main/java/com.terracode.blueharvest/listeners/displayValueListeners/MaxRPMDisplayListener.kt
@@ -1,11 +1,11 @@
-package com.terracode.blueharvest.listeners.configurationListeners
+package com.terracode.blueharvest.listeners.displayValueListeners
 
 import android.text.Editable
 import android.text.TextWatcher
 import com.terracode.blueharvest.ConfigurationSettingsActivity
 import com.terracode.blueharvest.utils.PreferenceManager
 
-class MaxHeightDisplayListener(private val activity: ConfigurationSettingsActivity) :
+class MaxRPMDisplayListener(private val activity: ConfigurationSettingsActivity) :
     TextWatcher {
 
     override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
@@ -22,7 +22,7 @@ class MaxHeightDisplayListener(private val activity: ConfigurationSettingsActivi
             if (input.isNotEmpty()) {
                 val value = input.toIntOrNull()
                 value?.let {
-                    PreferenceManager.setMaxHeightDisplayedInput(it)
+                    PreferenceManager.setMaxRPMDisplayedInput(it)
                 }
             }
         }
