+++ app/src/main/res/layouts/activities/layout/activity_configuration_settings.xml
@@ -93,7 +93,7 @@
                     android:id="@+id/operationParameterCard"
                     layout="@layout/safety_operational_settings_card"
                     android:layout_width="300dp"
-                    android:layout_height="wrap_content"
+                    android:layout_height="match_parent"
                     android:layout_marginStart="50dp"
                     android:layout_marginTop="20dp"
                     android:layout_marginEnd="25dp"
