+++ app/src/main/res/layouts/cards/layout/safety_operational_settings_card.xml
@@ -8,7 +8,7 @@
 
         <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="match_parent"
             android:background="#DADBDC"
             android:orientation="vertical">
 
@@ -102,6 +102,7 @@
                 android:background="@color/white"
                 android:ems="10"
                 android:gravity="center"
+                android:layout_marginTop="30dp"
                 android:layout_gravity="center_horizontal"
                 android:hint="@string/numberHint"
                 android:inputType="number"
@@ -125,6 +126,7 @@
                 android:background="@color/white"
                 android:ems="10"
                 android:gravity="center"
+                android:layout_marginTop="30dp"
                 android:layout_gravity="center_horizontal"
                 android:hint="@string/numberHint"
                 android:inputType="number"
